pyproject.toml
[tool.poetry]
name = "conversao-de-audios-teste"
version = "0.1.0"
description = ""
authors = ["Luciano Gomes <lucianogvda02@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
setuptools-rust = "^1.10.1"
llvmlite = "^0.43.0"
numba = "^0.60.0"
openai-whisper = {git = "https://github.com/openai/whisper.git"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
